+++ src/pygitgo/utils/platform.py
@@ -1,4 +1,4 @@
-from pygitgo.utils.cli_io import warning, info
+from pygitgo.utils.cli_io import warning, info, write
 import webbrowser
 import subprocess
 import platform
@@ -46,5 +46,5 @@ def open_url(url: str):
     if not opened:
         warning("Could not open browser automatically.")
         info("\nIf the browser did not open, visit this URL manually:")
-        print(f"\n  {url}\n")
+        write(f"\n  {url}\n")